+++ .eslintignore
@@ -1,2 +1,2 @@
 dist/
-lib/scripts/vendor/*
+lib/scripts/vendors/*
