+++ lib/scripts/main.es6.js
@@ -2,7 +2,7 @@
 require('../styles/executor.css');
 
 // Browser polyfills.
-require('./vendor/polyfills');
+require('./vendors/polyfills');
 
 // Defined default layout for JavaScript mode in Ace editor.
 require('../../bower_components/ace-builds/src-min/ace');
